SSL: Fixed memory leak in the plug-in
Unluckily, Rails' autoloading feature does not correctly garbage-collect
modules populated dynamically at runtime, or I'm not informing the loading
system correctly, either.

Anyway, regenerating the route helper method list, re-generating the
ssl_/plain_ helpers and re-filling the Module at each reload (read:
for each request in development mode) is a performance hog that slows
us down.

So, the solution is: an SSLRouteHelpers module outside the PM::SSL module,
that won't be garbage-collected because its name doesn't match the file
name in which is defined, populated one time and frozen afterwards.

The only thing that gets done on each reload!/request is to re-include
this module inside ApplicationController and friends, thus improving
performances at the expense of not having new route helpers defined
after the reload if you modify config/routes.rb.

Because we don't define new routes so often, it's an affordable drawback
we can live with. It's much worse to see our `thin` servers consume 2GB
of RAM, like it happened today on my and on Ferdinando's machines.

Oh, I forgot to say I suffer of logorrhoea/talkativeness :-P.