SSL finalization - ATTENTION: READ COMMIT MESSAGE
Because we run (both on pmdev.* and on our local machines) the application
on non-standard ports (8000 / 3000) you NEED TO UPDATE your config/settings.yml
file adding the following lines below the ":site_url:" setting

  :http_port: 3000
  :https_port: 443

and to remove the

  :site_url: localhost

line. Even if you don't use SSL today, because it will save headaches
tomorrow :-). Please remember that SSL is enable ONLY in production and
in test mode, so it won't disrupt your daily workflow in development mode.

Edo: on pmdev.* I configured Apache for SSL and added the relevant :http_port:
setting to the configuration file, to not break pmdev.