Filters: preserve all params when redirecting
Fixes an issue on Panmind where a redirect to /messages/compose/?to=XYZ
inside projects, loaded via HTTPS, redirected the user to the HTTP URI
but without the ?to=XYZ query string parameter.

So the ActionController params Hash is updated explicitly with SSL
settings, no more relying on AC doing it automagically.